requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openai>=1.14.0
python-dotenv>=1.0.0
tqdm>=4.66.0